{M103 PICK} [Fix UserAgentReduction Policy] (1) Make ForceMajorVersionToMinorPosition a class enum
Before we fix the issue with UserAgentReduction we should make the first
argument to GetProductAndVersion safer so that it won't be confused with
the new second argument.

This CL is part of a series:
(1) Make ForceMajorVersionToMinorPosition a class enum
(2) Fix GetProductAndVersion

(cherry picked from commit 55d0f400b0edc6171e000b82136e01b35c54f817)

Bug: 1330588
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3681901
Auto-Submit: Ari Chivukula <arichiv@chromium.org>
Commit-Queue: Ari Chivukula <arichiv@chromium.org>
Cr-Original-Commit-Position: refs/heads/main@{#1009892}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3686542
Cr-Commit-Position: refs/branch-heads/5060@{#633}
Cr-Branched-From: b83393d0f4038aeaf67f970a024d8101df7348d1-refs/heads/main@{#1002911}

+++ chrome/browser/chrome_content_browser_client.cc
@@ -1368,7 +1368,8 @@ void ChromeContentBrowserClient::RegisterProfilePrefs(
   registry->RegisterBooleanPref(prefs::kOriginAgentClusterDefaultEnabled, true);
   registry->RegisterIntegerPref(
       prefs::kForceMajorVersionToMinorPositionInUserAgent,
-      embedder_support::ForceMajorVersionToMinorPosition::kDefault);
+      static_cast<int>(
+          embedder_support::ForceMajorVersionToMinorPosition::kDefault));
   registry->RegisterBooleanPref(
       policy::policy_prefs::kIsolatedAppsDeveloperModeAllowed, true);
 }
